Move impl of `mazer version` from actions->mazer_version
Move the impl details of collecting the mazer version info
for `mazer version` to ansible_galaxy.mazer_version so that
other modules can use it (namely, ansible_galaxy.cli.galaxy
for logging info at cli startup)

Fixes #267
